Extract "add torrent params" serialization code
